# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "TeleFlow API Gateway",
    "version": "1.0.0"
  },
  "paths": {
...
    "/compose": {
      "post": {
        "tags": [
          "composer"
        ],
        "summary": "Compose",
        "operationId": "componer_draft",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ]
      }
    },
    "/drafts": {
      "get": {
        "tags": [
          "composer"
        ],
        "summary": "Drafts",
        "operationId": "listar_drafts",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ]
      }
    },
    "/drafts/{rest}": {
      "post": {
        "tags": [
          "composer"
        ],
        "summary": "Draft Ops",
        "operationId": "operar_draft",
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "parameters": [
          {
            "name": "rest",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Rest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "composer"
        ],
        "summary": "Draft Ops",
        "operationId": "obtener_draft",
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "parameters": [
          {
            "name": "rest",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Rest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/drafts/{draft_id}/source": {
      "patch": {
        "tags": [
          "composer"
        ],
        "summary": "Edit Draft",
        "description": "Corrige a mano la fuente de un borrador y la revalida.\n\nLleva `flows:deploy` y **no** `compose:write`, que es el scope del resto de las rutas de\nborradores. El corte no es \"qué recurso toca\" sino \"qué habilita\": aprobar despliega\n`draft.source`, así que quien puede reescribir esa columna decide qué código se va a\nejecutar. Es la misma responsabilidad que desplegar, y por eso el mismo permiso.\n\nConsecuencia buscada: una clave de analista puede pedir borradores y leerlos, y no puede\nreescribir lo que se despliega.",
        "operationId": "editar_draft",
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "parameters": [
          {
            "name": "draft_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Draft Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "APIKeyHeader": {
        "type": "apiKey",
        "in": "header",
        "name": "X-TeleFlow-API-Key"
      }
    }
  }
}